+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - New #111: Add `DateTime` and `DateTimeImmutable` support as time in log context (@vjik)
 - New #112: Add `Message::category()` method and `Message::DEFAULT_CATEGORY` constant, deprecate
   `CategoryFilter::DEFAULT` in favor it (@vjik)
+- Enh #113: Remove unnecessary `unset` call in `ContextProvider` (@vjik)
 
 ## 2.0.0 May 22, 2022