+++ CHANGELOG.md
@@ -9,14 +9,13 @@
 - Bug #98: Fix error on formatting trace, when it doesn't contain "file" and "line" (@vjik)
 - New #108: Support of nested values in message templates' variables, e. g. `{foo.bar}` (@vjik)
 - Bug #89: Fix error on parse messages, that contains variables that cannot cast to a string (@vjik)
-- New #109, #116: Add context providers (@vjik)
+- New #109, #113, #116: Add context providers (@vjik)
 - Chg #109: Deprecate `Logger` methods `setTraceLevel()` and `setExcludedTracePaths()` in favor of context provider
   usage (@vjik)
 - New #111: Add `DateTime` and `DateTimeImmutable` support as time in log context (@vjik)
 - New #112: Add `Message::category()` method and `Message::DEFAULT_CATEGORY` constant, deprecate
   `CategoryFilter::DEFAULT` in favor it (@vjik)
 - New #113: Add `Message::trace()` method (@vjik)
-- Enh #113: Remove unnecessary `unset` call in `ContextProvider` (@vjik)
 - New #114: Add `Message::time()` method (@vjik)
 - Chg #116: Deprecate methods `setCommonContext()` and `getCommonContext()` in `Target` class (@vjik)